requirements.txt
requests
flask
cryptography
Mastodon.py
gunicorn
